+++ src/config/config.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+	'listing_items_per_page' => 25
+];